Remove the tarot bot class for a functional approach
